add colab env check

+++ crop_from_video.py
@@ -9,7 +9,9 @@ import sys
 Video_Path = sys.argv[1]
 save_path = sys.argv[2]
 
-print(" Video_Path is:", Video_Path, " save_path:", save_path)
+colab_env = 'google.colab' in sys.modules
+
+print(" Video_Path is:", Video_Path, " save_path:", save_path, " colab_env: ", colab_env )
 
 #Video_Path = 'train/me2.mp4'
 video_Path = os.path.join(os.path.realpath('.'), Video_Path)
@@ -29,14 +31,16 @@ while (cap.isOpened()) and n<500:
     # n = n + 1
     if ret==True:
         #out.write(frame)
-        cv2.imshow('frame', frame)
-        if cv2.waitKey(1) & 0xFF == ord('q'):
-            break
+        if !colab_env:
+          cv2.imshow('frame', frame)
+          if cv2.waitKey(1) & 0xFF == ord('q'):
+              break
     n = n +1
     print(n)
 cap.release()
 out.release()
-cv2.destroyAllWindows()
+if !colab_env:
+  cv2.destroyAllWindows()
 
 # # Check if camera opened successfully
 # if (cap.isOpened() == False):
